step: add database set up checks for drizzle scripts

+++ drizzle-reset.sh
@@ -3,6 +3,12 @@
 source bgord-scripts/base.sh
 setup_base_config
 
+if ! test -d "infra/drizzle"
+then
+  info "Database not set up"
+  exit 0
+fi
+
 step_start "Database delete"
 rm -rf sqlite.db
 step_end "Database delete"
